requirements.txt
fastapi
uvicorn[standard]
python-multipart

torch
torchvision
transformers
pillow
opencv-python

numpy
scikit-image
scikit-learn

requests
aiofiles

pydantic
python-dotenv
webcolors
timm